requirements.txt
apache-airflow==2.8.1
pandas==2.1.4
psycopg2-binary==2.9.9
pyyaml==6.0.1
SQLAlchemy==1.4.51
jinja2==3.1.3
